ECO/paddle2.0/requirements.txt
Python>=3.7.4
PaddlePaddle>=2.0.0-rc1